requirements.txt
# Job Application Automator — Python Dependencies
# Install with: pip install -r requirements.txt

# Job scraping
jobspy>=0.31.0

# Anthropic Claude API
anthropic>=0.25.0

# Notion API
notion-client>=2.2.1

# DOCX generation/editing
python-docx>=1.1.0

# Optional: DOCX → PDF (requires Microsoft Word on Windows)
docx2pdf>=0.1.8

# Scheduling
schedule>=1.2.0

# Data handling
pandas>=2.0.0

# Utilities
requests>=2.31.0

# Local environment file loader
python-dotenv>=1.0.0
